preservative: change behaviour when asleep.
Get an early suspend hook from the hotplug driver to allow for
a drop the optimal frequency position. Drop opt_pos down to a level
where the bus speed is normal. Restore it when the device wakes (my
kernel uses mainly max bus speed when the device is not idle).

Allow scaling up to any fq even when asleep to ensure the dialer
app loads quickly.

Prevent touch events from boosting the fq when asleep.